# Add AppConfig client performance test