--- a/src/SDKs/Azure.ApplicationModel.Configuration/data-plane/Azure.ApplicationModel.Configuration.Performance/Azure.ApplicationModel.Configuration.Performance.csproj
+++ b/src/SDKs/Azure.ApplicationModel.Configuration/data-plane/Azure.ApplicationModel.Configuration.Performance/Azure.ApplicationModel.Configuration.Performance.csproj
@@ -1,0 +1,20 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>$(RequiredTargetFrameworks)</TargetFrameworks>
+    <OutputType>Exe</OutputType>
+    <ImportDefaultReferences>false</ImportDefaultReferences>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotNet" Version="0.11.5" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Azure.ApplicationModel.Configuration\Azure.Configuration.csproj" />
+  </ItemGroup>
+
+    <!-- Import the Azure.Core project -->
+  <Import Project="$(MSBuildThisFileDirectory)..\..\..\Azure.Core\data-plane\Azure.Core\Azure.Core.props" />
+
+</Project>