--- a/src/SDKs/Azure.ApplicationModel.Configuration/data-plane/Azure.ApplicationModel.Configuration.Performance/ConfigurationClientBenchmark.cs
+++ b/src/SDKs/Azure.ApplicationModel.Configuration/data-plane/Azure.ApplicationModel.Configuration.Performance/ConfigurationClientBenchmark.cs
@@ -19,7 +19,7 @@ namespace Azure.ApplicationModel.Configuration.Performance
 
         private static readonly MockClientHandler GetResponseMessage = new MockClientHandler(request => new HttpResponseMessage((HttpStatusCode)200)
         {
-            Content = new ByteArrayContent(_getBytes)
+            Content = new ByteArrayContent(GetResponseBytes)
         });
 
         private static readonly ConfigurationClient ConfigurationClient = new ConfigurationClient(ConnectionString, new ConfigurationClientOptions()
@@ -27,7 +27,7 @@ namespace Azure.ApplicationModel.Configuration.Performance
             Transport = new HttpClientTransport(new HttpClient(GetResponseMessage))
         });
 
-        private static byte[] _getBytes = Encoding.UTF8.GetBytes(@"{
+        private static readonly byte[] GetResponseBytes = Encoding.UTF8.GetBytes(@"{
   ""etag"": ""4f6dd610dd5e4deebc7fbaef685fb903"",
   ""key"": ""key"",
   ""label"": ""label"",
@@ -42,7 +42,7 @@ namespace Azure.ApplicationModel.Configuration.Performance
 }");
 
         [Benchmark]
-        public async Task GetSecretBenchmarkCore()
+        public async Task GetAsync()
         {
             await ConfigurationClient.GetAsync("key");
         }