--- a/src/SDKs/Azure.ApplicationModel.Configuration/data-plane/Azure.ApplicationModel.Configuration.sln
+++ b/src/SDKs/Azure.ApplicationModel.Configuration/data-plane/Azure.ApplicationModel.Configuration.sln
@@ -5,11 +5,13 @@ VisualStudioVersion = 15.0.28307.231
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Azure.Configuration", "Azure.ApplicationModel.Configuration\Azure.Configuration.csproj", "{25CE0676-8F2C-497B-8FB5-7DA41A31F6F9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.Configuration.Tests", "Azure.ApplicationModel.Configuration.Tests\Azure.Configuration.Tests.csproj", "{AAA761F4-D037-46AC-BF7A-496C15310F66}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Azure.Configuration.Tests", "Azure.ApplicationModel.Configuration.Tests\Azure.Configuration.Tests.csproj", "{AAA761F4-D037-46AC-BF7A-496C15310F66}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Azure.Core", "..\..\Azure.Core\data-plane\Azure.Core\Azure.Core.csproj", "{019A28FC-EFEA-4431-872C-AB8F9C0049BF}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Azure.Core.Tests", "..\..\Azure.Core\data-plane\Azure.Core.Tests\Azure.Core.Tests.csproj", "{84491222-6C36-4FA7-BBAE-1FA804129151}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Azure.ApplicationModel.Configuration.Performance", "Azure.ApplicationModel.Configuration.Performance\Azure.ApplicationModel.Configuration.Performance.csproj", "{2F3CC223-2357-4D24-BCE7-92835C62A201}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -33,6 +35,10 @@ Global
 		{84491222-6C36-4FA7-BBAE-1FA804129151}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{84491222-6C36-4FA7-BBAE-1FA804129151}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{84491222-6C36-4FA7-BBAE-1FA804129151}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2F3CC223-2357-4D24-BCE7-92835C62A201}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2F3CC223-2357-4D24-BCE7-92835C62A201}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2F3CC223-2357-4D24-BCE7-92835C62A201}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2F3CC223-2357-4D24-BCE7-92835C62A201}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE