--- a/src/SDKs/Azure.Core/data-plane/Azure.Core/Pipeline/HttpClientTransport.cs
+++ b/src/SDKs/Azure.Core/data-plane/Azure.Core/Pipeline/HttpClientTransport.cs
@@ -145,30 +145,34 @@ namespace Azure.Core.Pipeline
 
             public HttpRequestMessage BuildRequestMessage(CancellationToken cancellation)
             {
-                if (_wasSent == false)
+                HttpRequestMessage currentRequest;
+                if (_wasSent)
                 {
+                    // A copy of a message needs to be made because HttpClient does not allow sending the same message twice,
+                    // and so the retry logic fails.
+                    currentRequest = new HttpRequestMessage(_requestMessage.Method, UriBuilder.Uri);
+                    CopyHeaders(_requestMessage.Headers, currentRequest.Headers);
+
+                }
+                else
+                {
+                    currentRequest = _requestMessage;
                     _wasSent = true;
-                    _requestMessage.RequestUri = UriBuilder.Uri;
-                    return _requestMessage;
                 }
 
-                // A copy of a message needs to be made because HttpClient does not allow sending the same message twice,
-                // and so the retry logic fails.
-                var request = new HttpRequestMessage(_requestMessage.Method, UriBuilder.Uri);
-
-                CopyHeaders(_requestMessage.Headers, request.Headers);
+                _requestMessage.RequestUri = UriBuilder.Uri;
 
                 if (_requestContent?.PipelineContent != null)
                 {
-                    request.Content = new PipelineContentAdapter()
+                    currentRequest.Content = new PipelineContentAdapter()
                     {
                         CancellationToken = cancellation,
                         PipelineContent = _requestContent.PipelineContent
                     };
-                    CopyHeaders(_requestContent.Headers, request.Content.Headers);
+                    CopyHeaders(_requestContent.Headers, currentRequest.Content.Headers);
                 }
 
-                return request;
+                return currentRequest;
             }
 
             public override void Dispose()