--- a/src/SDKs/Azure.Core/data-plane/Azure.Core/Pipeline/Policies/BufferResponsePolicy.cs
+++ b/src/SDKs/Azure.Core/data-plane/Azure.Core/Pipeline/Policies/BufferResponsePolicy.cs
@@ -19,7 +19,7 @@ namespace Azure.Core.Pipeline.Policies
         {
             await ProcessNextAsync(pipeline, message);
 
-            if (message.Response.ResponseContentStream != null)
+            if (message.Response.ResponseContentStream != null && !message.Response.ResponseContentStream.CanSeek)
             {
                 Stream responseContentStream = message.Response.ResponseContentStream;
                 var bufferedStream = new MemoryStream();