I've added a microbenchmark for AppConfig client GetAsync overhead

Did a couple simple perf optimization in our version:

1. Avoid re-creating HttpRequestMessage when sending only once
2. Avoid examining request/response if logging is disabled
3. Avoid buffering if response stream is already buffered

Results compared to https://github.com/Azure/Azconfig-DotnetSdk (GetSecretBenchmarkAz) on .net core 2.1

```

|                 Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------- |---------:|----------:|----------:|-------:|------:|------:|----------:|
|   GetSecretBenchmarkAz | 33.09 us | 0.6594 us | 0.7593 us | 1.9531 |     - |     - |  12.32 KB |
| GetSecretBenchmarkCore | 25.80 us | 0.5282 us | 0.6487 us | 1.0681 |     - |     - |   6.71 KB |

```

For net461 and netcoreapp2.0 iteration time is similar between both libraries but allocation for Azure.Core one is almost 2x less.