@pakrym 2019-04-23
> Do we have end-to-end perf tests of AppConfig against the live Azure service

Not that I know of, the service team may have some.

> or are there plans to write them?

There are no current plans.